requirements.txt
numpy
dpdata
pydflow>=1.0.20
dargs
ase
pydflow
phonopy
phonoLAMMPS